[TLE] Add WGMMA pipeline mode selector
