+++ include/triton/Tools/Sys/GetEnv.hpp
@@ -45,6 +45,7 @@ inline const std::set<std::string> CACHE_INVALIDATING_ENV_VARS = {
     "TRITON_F32_DEFAULT",
     "TRITON_PREFER_TMEM_16x256_LAYOUT",
     "TRITON_ENABLE_EXPERIMENTAL_CONSAN",
+    "TLE_WGMMA_PIPELINE_MODE",
     // clang-format on
 };
