+++ python/src/passes.cc
@@ -66,7 +66,15 @@ void init_triton_passes_ttgpuir(py::module &&m) {
   ADD_PASS_OPTION_WRAPPER_1("add_assign_latencies",
                             createTritonGPUAssignLatencies, int);
   ADD_PASS_WRAPPER_0("add_schedule_loops", createTritonGPUScheduleLoops);
-  ADD_PASS_OPTION_WRAPPER_2("add_pipeline", createTritonGPUPipeline, int, bool);
+  m.def("add_pipeline",
+        [](mlir::PassManager &pm, int numStages, bool dumpIntermediateSteps,
+           const std::string &tleWgmmaPipelineMode) {
+          pm.addPass(createTritonGPUPipeline(
+              {numStages, dumpIntermediateSteps, tleWgmmaPipelineMode}));
+        },
+        py::arg("pm"), py::arg("num_stages"),
+        py::arg("dump_intermediate_steps"),
+        py::arg("tle_wgmma_pipeline_mode") = "compiler_auto");
   ADD_PASS_OPTION_WRAPPER_1("add_warp_specialize",
                             createTritonGPUAutomaticWarpSpecialization, int);
   ADD_PASS_WRAPPER_0("add_prefetch", createTritonGPUPrefetch);
